requirements.txt.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
aiosmtplib
celery
redis
python-dotenv
flask
